Turn off the `Symbol#to_proc` transform by default
When you convert something like:

```ruby
[1, 2, 3].map { |i| i.to_s }
```

into:

```ruby
[1, 2, 3].map(&:to_s)
```

it normally is completely fine. However, it technically changes the semantics of what's being passed by adding a method call (`Symbol#to_proc`) and changing the arity of the block being passed from `1` to `-1`. In some places this results in issues where reflection is being used. In other places it becomes problematic because they're looking for procs that aren't checking arity. Either way, I don't think we can justify changing the semantics this much.
