+++ test/js/blocks.test.js
@@ -123,116 +123,4 @@ describe("blocks", () => {
     test("does not split up args inside pipes", () =>
       expect(`loop do |${long} = 1, a${long} = 2|\nend`).toMatchFormat());
   });
-
-  describe("to_proc transform", () => {
-    test("basic inline", () =>
-      expect("loop { |i| i.to_s }").toChangeFormat("loop(&:to_s)"));
-
-    test("basic inline with option turned off", () =>
-      expect("loop { |i| i.to_s }").toMatchFormat({ toProcTransform: false }));
-
-    test("basic multi-line", () => {
-      const content = ruby(`
-        list.each do |node|
-          node.print
-        end
-      `);
-
-      return expect(content).toChangeFormat("list.each(&:print)");
-    });
-
-    test.skip("multi-line with comment", () => {
-      const content = ruby(`
-        foo.each do |bar|
-          # comment
-          bar.baz
-        end
-      `);
-
-      return expect(content).toMatchFormat();
-    });
-
-    test("happens for command nodes", () => {
-      const content = ruby(`
-        command 'foo' do |bar|
-          bar.to_s
-        end
-      `);
-
-      return expect(content).toChangeFormat("command 'foo', &:to_s");
-    });
-
-    test("happens for command call nodes", () => {
-      const content = ruby(`
-        command.call 'foo' do |bar|
-          bar.to_s
-        end
-      `);
-
-      return expect(content).toChangeFormat("command.call 'foo', &:to_s");
-    });
-
-    test("with args and parens", () =>
-      expect("foo(bar) { |baz| baz.to_i }").toChangeFormat("foo(bar, &:to_i)"));
-
-    test("with commands", () => {
-      const content = ruby(`
-        command bar do |baz|
-          baz.to_i
-        end
-      `);
-
-      return expect(content).toChangeFormat("command bar, &:to_i");
-    });
-
-    test("with command calls", () => {
-      const content = ruby(`
-        command.call bar do |baz|
-          baz.to_i
-        end
-      `);
-
-      return expect(content).toChangeFormat("command.call bar, &:to_i");
-    });
-
-    test("does not happen when there are multiple lines", () => {
-      const content = ruby(`
-        loop do |i|
-          i.to_s
-          i.next
-        end
-      `);
-
-      return expect(content).toMatchFormat();
-    });
-
-    test("does not happen when there are args to the method call", () =>
-      expect("loop { |i| i.to_s(:db) }").toMatchFormat());
-
-    test("does not happen when there are multiple args", () =>
-      expect("loop { |i, j| i.to_s }").toMatchFormat());
-
-    test("does not duplicate when inside of an aref node", () =>
-      expect("foo[:bar].each { |baz| baz.to_s }").toChangeFormat(
-        "foo[:bar].each(&:to_s)"
-      ));
-
-    describe.each(["if", "unless"])(
-      "does not transform when used inside hash with %s",
-      (keyword) => {
-        test(`hash literal with :${keyword} key`, () =>
-          expect(`{ ${keyword}: ->(foo) { foo.to_s } }`).toMatchFormat());
-        test(`hash literal with hashrocket :${keyword} key`, () =>
-          expect(`{ :${keyword} => ->(foo) { foo.to_s } }`).toMatchFormat({
-            preferHashLabels: false
-          }));
-        test(`method arguments with :${keyword} key`, () =>
-          expect(`bar ${keyword}: ->(foo) { foo.to_s }`).toMatchFormat());
-        test(`method arguments with hashrocket :${keyword} key`, () =>
-          expect(`bar :${keyword} => ->(foo) { foo.to_s }`).toMatchFormat({
-            preferHashLabels: false
-          }));
-      }
-    );
-  });
 });
